Add balance endpoint plus minor fixes to Sendgrid automation

+++ src/routes/controllers/ManagerControllers/flagDiscussion.ts
@@ -6,28 +6,33 @@ import CustomersModel, { Customers } from '../../../models/Customers';
 export const flagDiscussion = async (req: Request, res: Response) => {
   try {
     const postID = req.params.discussionPostId;
-    var discussionLink = 'http://localhost:5000/api/discussions/' + postID;
-    const discussionPost = await axios.get(discussionLink);
-    var customerID = discussionPost.data.messageFrom; // get customerID of person who created the discussion post
+    const discussionPost: Discussions | null = await DiscussionsModel.findById(postID);
 
+    if (discussionPost) {
+      var customerID = discussionPost.messageFrom; // get customerID of person who created the discussion post
+      const customer: Customers | null = await CustomersModel.findById(customerID);
 
-    var customerProfile = 'http://localhost:5000/api/customers/' + customerID;
-    const customer = await axios.get(customerProfile);
-    var numOfWarnings = customer.data.warnings; // get number of warnings for customer
-    var vipStatus = customer.data.isVIP; // get VIP status for customer
+      if (customer) {
+        var numOfWarnings = customer.warnings; // get number of warnings for customer
+        var vipStatus = customer.isVIP; // get VIP status for customer
 
-    if (vipStatus == false && numOfWarnings < 3) // if numOfWarnings < 3, increment numOfWarnings
-      await CustomersModel.findByIdAndUpdate(customerID, { warnings: ++numOfWarnings });
-    else if (vipStatus == false && numOfWarnings >= 3) // if numOfWarnings >= 3, blacklist customer
-      await CustomersModel.findByIdAndUpdate(customerID, { isBlacklisted: true });
-    else if (vipStatus == true && numOfWarnings < 2) // if VIP and numOfWarnings < 2, increment numOfWarnings
-      await CustomersModel.findByIdAndUpdate(customerID, { warnings: ++numOfWarnings });
-    else if (vipStatus == true && numOfWarnings >= 2) // if VIP and numOfWarnings >= 2, get rid of VIP staus
-      await CustomersModel.findByIdAndUpdate(customerID, { isVIP: false, warnings: 0 });
+        if (vipStatus == false && numOfWarnings < 3) // if numOfWarnings < 3, increment numOfWarnings
+          await CustomersModel.findByIdAndUpdate(customerID, { warnings: ++numOfWarnings });
+        else if (vipStatus == false && numOfWarnings >= 3) // if numOfWarnings >= 3, blacklist customer
+          await CustomersModel.findByIdAndUpdate(customerID, { isBlacklisted: true });
+        else if (vipStatus == true && numOfWarnings < 2) // if VIP and numOfWarnings < 2, increment numOfWarnings
+          await CustomersModel.findByIdAndUpdate(customerID, { warnings: ++numOfWarnings });
+        else if (vipStatus == true && numOfWarnings >= 2) // if VIP and numOfWarnings >= 2, get rid of VIP staus
+          await CustomersModel.findByIdAndUpdate(customerID, { isVIP: false, warnings: 0 });
 
-    await DiscussionsModel.findByIdAndDelete(postID); // delete post
-
-    res.json({ msg: 'Discussion post deleted!' });
+        await DiscussionsModel.findByIdAndDelete(postID); // delete post
+        res.json({ msg: 'Discussion post deleted!' });
+      }
+      else
+        res.status(404).json({ msg: 'Customer not found' });      
+    }
+    else
+      res.status(404).json({ msg: 'Discussion post not found' });
   } catch (error) {
     console.error(error.message);
     res.status(500).json({ msg: 'Server error' });
